MessageTab - Attempt to create inbox functionality
git-svn-id: https://svn.webgres.cz/svn/poptavka@849 61682b0c-635e-44ba-a85b-d27e7dc34134

+++ poptavka-web/src/main/java/cz/poptavka/sample/client/user/messages/MessagesModuleMessageHandler.java
@@ -92,7 +92,7 @@ public class MessagesModuleMessageHandler extends BaseEventHandler<MessagesModul
     }
 
     public void onRequestConversation(Long threadRootId, Long subRootId) {
-        messageService.getConversationMessages((long) threadRootId, (long) subRootId,
+        messageService.getConversationMessages(threadRootId, subRootId,
                 new AsyncCallback<ArrayList<MessageDetail>>() {
 
                     @Override
